Handle electricity sublayer with GPU data filters
The previous approach based on drawing transparent lines didn't work
for pickinng - the hidden lines were's still highlighted on hover.
The new approach uses the on-GPU data filtering extension to deck.gl
which truly hides the hidden layers. That way, it's not necessary to
handle data loading and filtering outside of the layer descriptor.