move fudge factor from frontend to backend
there is a problem with some data in the database, specifically i think the
adaptation_cost_benefit.avoided_eael_* values. these values are multiplied by
eael_days, which is sent from the FE. the avoided_eael_* values are corrected
with a fudge factor on eael_days.

this fudge factor is now applied in validation on receiving the value from the
FE

this fix should be removed asap and the data in the database corrected