Fix tiled selection layer for transparent polygons
The selectable MVT layer mechanism had a bug for polygon data.
When a transparent fill was set for the selection style,
tile layers were visible in the data. This was because
an `extensions` prop set by MVTLayer was being overwritten
by the `tileSelectionLayer` and therefore the tile geometries
were not cropped properly.